TO-DO-FastApi-localdb-authorization/requirements.txt
fastapi[standard]
uvicorn[standard]
sqlalchemy
pymysql
cryptography
passlib
python-jose[cryptography]
pyjwt
bcrypt